Resolve several PHASE2_BUILD.md uncertainties via deeper source reading
Fully resolved (implementation, not just .i declarations, read this time):

- sendAccountTextMessage's flag is a bitmask (bit 0 = "only if peer
  connected"), confirmed in configurationmanager.cpp. sendTextMessage now
  takes onlyIfConnected: Boolean instead of a bare int.
- Conversation sendMessage's flag is an unrelated action selector (0=send/
  reply, 1=edit, 2=react), confirmed in conversation_interface.cpp. Modeled
  as RealConversationMessageAction; sendConversationMessage takes it instead
  of a bare int.
- registerName's scheme/password are the account's own local archive
  password (needed to sign the registration with the account's private
  key), not name-server credentials, confirmed in
  archive_account_manager.cpp. Empty defaults are genuinely correct for
  unprotected accounts (fileutils::ARCHIVE_AUTH_SCHEME_NONE is literally
  ""), not a placeholder -- now threaded through as real parameters instead
  of hardcoded.
- Unsigned integer widening: SWIG's Java target widens unsigned char/short/
  int to short/int/long respectively (uint64_t would go to BigInteger,
  which is exactly why this project's own jni_interface.i has
  `%apply int64_t { uint64_t }` with a comment saying so). Applied
  throughout wherever a raw uint32_t appeared without a matching override:
  swarmLoaded/messagesFound's id, onConversationError's code,
  loadConversation/loadSwarmUntil/countInteractions/searchConversation's
  return values, searchConversation's maxResult, queryPeerServices's return
  value and peerServicesReceived's requestId, and
  DataTransferEventCode/DataTransferError -- all corrected from Int to Long.

PHASE2_BUILD.md's uncertainty list is split into "Resolved this pass" and
"Still open" so the remaining genuinely-unverified items (VectMap/List,
fileTransferInfo's OUTPUT array type, getCameraInfo's vector pointers,
SwarmMessage.status, size_t) aren't lost among the resolved ones.

+++ daemon/src/main/java/org/meshly/app/daemon/JamiCallbackAdapter.kt
@@ -196,12 +196,15 @@ internal class MeshlyDataTransferCallback(
     private val events: MutableSharedFlow<RealJamiEvent>
 ) : DataTransferCallback() {
 
+    // `eventCode` is `uint32_t` -- `Long`, per the widening convention documented on
+    // RealDataTransferEventCode (this project's own uint64_t->BigInteger-avoidance override is
+    // corroborating evidence SWIG's stock unsigned-widening defaults apply here).
     override fun dataTransferEvent(
         accountId: String,
         conversationId: String,
         interactionId: String,
         fileId: String,
-        eventCode: Int
+        eventCode: Long
     ) {
         events.tryEmit(
             RealJamiEvent.DataTransferEvent(
@@ -276,12 +279,13 @@ internal class MeshlyConversationCallback(
     private val events: MutableSharedFlow<RealJamiEvent>
 ) : ConversationCallback() {
 
-    override fun swarmLoaded(id: Int, accountId: String, conversationId: String, messages: SwarmMessageVect) {
+    // `id` is `uint32_t` in both methods below -- `Long`, see RealDataTransferEventCode's doc.
+    override fun swarmLoaded(id: Long, accountId: String, conversationId: String, messages: SwarmMessageVect) {
         val mapped = (0 until messages.size()).map { RealSwarmMessage.fromSwarmMessage(messages[it]) }
         events.tryEmit(RealJamiEvent.SwarmLoaded(id, accountId, conversationId, mapped))
     }
 
-    override fun messagesFound(id: Int, accountId: String, conversationId: String, messages: VectMap) {
+    override fun messagesFound(id: Long, accountId: String, conversationId: String, messages: VectMap) {
         events.tryEmit(RealJamiEvent.MessagesFound(id, accountId, conversationId, messages))
     }
 
@@ -346,7 +350,8 @@ internal class MeshlyConversationCallback(
         )
     }
 
-    override fun onConversationError(accountId: String, conversationId: String, code: Int, what: String) {
+    // `code` is `uint32_t` -- `Long`, see RealDataTransferEventCode's doc.
+    override fun onConversationError(accountId: String, conversationId: String, code: Long, what: String) {
         events.tryEmit(RealJamiEvent.ConversationError(accountId, conversationId, code, what))
     }
 
@@ -369,13 +374,15 @@ internal class MeshlyNetworkServiceCallback(
     private val events: MutableSharedFlow<RealJamiEvent>
 ) : NetworkServiceCallback() {
 
-    // `requestId` is `uint32_t` and `localPort` is `uint16_t` on the C++ side; no explicit `%apply`
-    // override for either exists in jni_interface.i (unlike uint64_t/time_t), so both are assumed
-    // to fall back to SWIG's un-overridden defaults (`int` for unsigned int, `short` for unsigned
-    // short) here as `Int` -- not confirmed against a real generated build, and `localPort` in
-    // particular could plausibly come out as `Short` instead.
+    // `requestId` is `uint32_t` (-> `Long`) and `localPort` below is `uint16_t` (-> `Int`); no
+    // explicit `%apply` override for either exists in jni_interface.i (unlike uint64_t/time_t),
+    // so both fall back to SWIG's documented default unsigned-widening rule: unsigned short ->
+    // int, unsigned int -> long, unsigned long/uint64_t -> BigInteger (this project's own
+    // `%apply int64_t { uint64_t }` override, commented "Avoid uint64_t to be converted to
+    // BigInteger", is direct evidence it follows exactly that convention). Reasonably confident,
+    // not confirmed against a real generated build.
     override fun peerServicesReceived(
-        requestId: Int,
+        requestId: Long,
         accountId: String,
         peerId: String,
         status: Int,
